order

+++ src/main/java/com/graduate/hou/service/impl/OrderServiceImpl.java
@@ -1,10 +1,7 @@
 package com.graduate.hou.service.impl;
 
 import com.graduate.hou.dto.request.OrderDTO;
-import com.graduate.hou.entity.Address;
-import com.graduate.hou.entity.Order;
-import com.graduate.hou.entity.Payment;
-import com.graduate.hou.entity.User;
+import com.graduate.hou.entity.*;
 import com.graduate.hou.repository.AddressRepository;
 import com.graduate.hou.repository.OrderRepository;
 import com.graduate.hou.repository.PaymentRepository;
@@ -36,6 +33,11 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
+    public Order findOrderById(Long id) {
+         return orderRepository.findById(id).get();
+    }
+
+    @Override
     public Order createOrder(OrderDTO orderDTO) {
         User user = usersRepository.findById(orderDTO.getUserId())
                 .orElseThrow(()-> new RuntimeException("Chưa đăng nhập"));
@@ -50,8 +52,6 @@ public class OrderServiceImpl implements OrderService {
                 .user(user)
                 .totalAmount(orderDTO.getTotalAmount())
                 .status(orderDTO.getStatus())
-                // .createdAt(orderDTO.getCreatedAt())
-                // .updatedAt(orderDTO.getUpdatedAt())
                 .payment(payment)
                 .address(address)
                 .build();
@@ -60,7 +60,7 @@ public class OrderServiceImpl implements OrderService {
 
     @SuppressWarnings("static-access")
     @Override
-    public Order updateOrder(Long id, OrderDTO orderDTO) {
+    public boolean updateOrder(Long id, OrderDTO orderDTO) {
         Optional<Order> optionalOrder = orderRepository.findById(id);
 
         User user = usersRepository.findById(orderDTO.getUserId())
@@ -72,21 +72,35 @@ public class OrderServiceImpl implements OrderService {
         Address address = addressRepository.findById(orderDTO.getAddressId())
                 .orElseThrow(()-> new RuntimeException("Chưa chọn địa chỉ"));
 
-        Order order = optionalOrder.get().builder()
-                .user(user)
-                .totalAmount(orderDTO.getTotalAmount())
-                .status(orderDTO.getStatus())
-                // .createdAt(orderDTO.getCreatedAt())
-                // .updatedAt(orderDTO.getUpdatedAt())
-                .payment(payment)
-                .address(address)
-                .build();
-        return orderRepository.save(order);
+        Order order = optionalOrder.get();
+        if(order != null){
+            order.setUser(user);
+            order.setTotalAmount(orderDTO.getTotalAmount());
+            order.setStatus(orderDTO.getStatus());
+            order.setPayment(payment);
+            order.setAddress(address);
+            try {
+                orderRepository.save(order);
+                return true;
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return false;
     }
 
 
     @Override
-    public void deleteOrder(Long id) {
-        orderRepository.deleteById(id);
+    public boolean deleteOrder(Long id) {
+        Order order = orderRepository.findById(id).get();
+        if(order != null){
+            try {
+                orderRepository.deleteById(id);
+                return true;
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+        }
+        return true;
     }
 }
